chore(api): split document cmd file
Ref. metriport/metriport-internal#515
